refactor: Outsource home page component and button
