+++ 9-simple-exercises/state props/src/components/Button.jsx
@@ -1,0 +1,5 @@
+function Button({ onClick }) {
+    return <button onClick={onClick}>Click me</button>
+}
+
+export default Button
